requirements.txt
appdirs==1.4.3
args==0.1.0
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
clint==0.5.1
colorama==0.4.1
funcparserlib==0.3.6
html2text==2018.1.9
hy==0.16.0
Markdown==3.1
more-itertools==7.0.0
pluggy==0.9.0
py==1.8.0
pyreadline==2.1
pytest==4.4.1
rply==0.7.7
six==1.12.0
pyyaml==5.1
soupsieve==1.9.1
